+++ host_agent/src/libraries/runtime/win32/Api_ABC.h
@@ -57,7 +57,7 @@ struct Api_ABC : public boost::noncopyable
     virtual bool              TerminateProcess     ( void* hProcess, unsigned uExitCode ) const = 0;
     virtual std::wstring      GetModuleFilename    () const = 0;
     virtual ProcessDescriptor MakeProcess          ( const wchar_t* app, wchar_t* args, const wchar_t* run, const wchar_t* log ) const = 0;
-    virtual unsigned long     GetCurrentProcessIdentifier() const = 0;
+    virtual int               GetCurrentProcessIdentifier() const = 0;
     //@}
 };
 }
